+++ app/src/lib.rs
@@ -231,6 +231,10 @@ pub fn configure_routes(cfg: &mut web::ServiceConfig) {
                             ),
                     ),
                 )
+                .service(web::scope("/protocol").route(
+                    "/metrics",
+                    web::get().to(api::protocol::get_protocol_metrics),
+                ))
                 .service(
                     web::scope("/evm")
                         .configure(api::creator::configure)